Refactor the app.rb
