+++ app.rb
@@ -1,10 +1,14 @@
 require_relative 'book'
 require_relative 'person'
 require_relative 'rental'
-require_relative 'student'
-require_relative 'teacher'
+require_relative 'creation'
+require_relative 'list'
 
 class App
+  include Creator
+  include Lister
+  attr_accessor :books, :people, :rentals
+
   def initialize
     @books = []
     @rentals = []
